refactor(lint): extract board methods to ConnectionRuntime+Boards.swift
Fixes SwiftLint type_body_length error — class body was 2566 lines (limit
2500). Board Models and Board Messages sections moved to an @MainActor
extension in a new file; no logic changes.
